+++ examples/example/settings.py
@@ -1,12 +1,11 @@
 # Django settings for example project.
 #
 # These settings are pretty basic.
-# The only relevent settings are the defaults are:
+# The only relevant settings are the defaults are:
 # - INSTALLED_APPS
 # - COMMENTS_APP
 #
 # And the following are configured to sane defaults:
-# - TEMPLATE_CONTEXT_PROCESSORS
 # - STATIC_ROOT / STATIC_URL
 # - MEDIA_ROOT / MEDIA_URL
 #
@@ -14,7 +13,6 @@ import os
 import django
 
 DEBUG = True
-TEMPLATE_DEBUG = DEBUG
 
 ADMINS = (
     # ('Your Name', 'your_email@domain.com'),
@@ -26,7 +24,7 @@ ADMINS = (
 MANAGERS = ADMINS
 
 
-## -- Server specific settings
+# -- Server specific settings
 
 DATABASES = {
     'default': {
@@ -45,7 +43,7 @@ SECRET_KEY = 'sfm=0t(!sqi&!y%66+e+#4m$1o&l%(l(w#vz$=_0c$5+#m*9yk'
 SITE_ID = 1
 
 
-## -- Internal Django config
+# -- Internal Django config
 
 # Language codes
 TIME_ZONE = 'America/Chicago'
@@ -55,7 +53,7 @@ USE_L10N = True
 USE_TZ = True
 
 # Paths, using autodetection
-PROJECT_DIR  = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
+PROJECT_DIR = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
 
 MEDIA_ROOT = os.path.join(PROJECT_DIR, 'media')
 MEDIA_URL = '/media/'
@@ -69,22 +67,22 @@ ROOT_URLCONF = 'example.urls'
 ADMIN_MEDIA_PREFIX = '/static/admin/'
 
 
-## -- Plugin components
+# -- Plugin components
 
-TEMPLATE_LOADERS = (
-    'django.template.loaders.filesystem.Loader',
-    'django.template.loaders.app_directories.Loader',
-)
-
-TEMPLATE_CONTEXT_PROCESSORS = (
-    'django.core.context_processors.debug',
-    'django.core.context_processors.i18n',
-    'django.core.context_processors.media',
-    'django.core.context_processors.request',    # Very useful to have, not found in default Django setup.
-    'django.core.context_processors.static',
-    'django.contrib.auth.context_processors.auth',
-    'django.contrib.messages.context_processors.messages',
-)
+TEMPLATES = [
+    {
+        'BACKEND': 'django.template.backends.django.DjangoTemplates',
+        'APP_DIRS': True,
+        'OPTIONS': {
+            'context_processors': [
+                'django.contrib.auth.context_processors.auth',
+                'django.template.context_processors.request',
+                'django.template.context_processors.static',
+                'django.contrib.messages.context_processors.messages',
+            ]
+        },
+    },
+]
 
 MIDDLEWARE_CLASSES = (
     'django.middleware.common.CommonMiddleware',
@@ -93,16 +91,14 @@ MIDDLEWARE_CLASSES = (
     'django.contrib.auth.middleware.AuthenticationMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
 )
+MIDDLEWARE = MIDDLEWARE_CLASSES  # Django 2.0
 
 STATICFILES_FINDERS = (
     'django.contrib.staticfiles.finders.FileSystemFinder',
     'django.contrib.staticfiles.finders.AppDirectoriesFinder',
-#    'django.contrib.staticfiles.finders.DefaultStorageFinder',
 )
 
 STATICFILES_DIRS = ()
-
-TEMPLATE_DIRS = ()
 
 INSTALLED_APPS = (
     'django.contrib.auth',
@@ -118,12 +114,12 @@ INSTALLED_APPS = (
     'threadedcomments',
 )
 
-if django.VERSION >= (1,7):
+if django.VERSION >= (1, 8):
     INSTALLED_APPS += ('django_comments',)
 else:
     INSTALLED_APPS += ('django.contrib.comments',)
 
 
-## --- App settings
+# --- App settings
 
 COMMENTS_APP = 'threadedcomments'