+++ threadedcomments/admin.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
 from django.utils.translation import ugettext_lazy as _
+
 from .compat import BASE_APP
 from .models import ThreadedComment
 
@@ -19,17 +20,17 @@ else:
 class ThreadedCommentsAdmin(CommentsAdmin):
     fieldsets = (
         (None,
-           {'fields': ('content_type', 'object_pk', 'site')}
-        ),
+         {'fields': ('content_type', 'object_pk', 'site')}
+         ),
         (_('Content'),
-           {'fields': ('user', 'user_name', 'user_email', 'user_url', 'title', 'comment')}
-        ),
+         {'fields': ('user', 'user_name', 'user_email', 'user_url', 'title', 'comment')}
+         ),
         (_('Hierarchy'),
-           {'fields': ('parent',)}
-        ),
+         {'fields': ('parent',)}
+         ),
         (_('Metadata'),
-           {'fields': ('submit_date', 'ip_address', 'is_public', 'is_removed')}
-        ),
+         {'fields': ('submit_date', 'ip_address', 'is_public', 'is_removed')}
+         ),
     )
 
     list_display = ('name', 'title', 'content_type', 'object_pk', 'parent',
@@ -39,4 +40,3 @@ class ThreadedCommentsAdmin(CommentsAdmin):
     raw_id_fields = ("parent",)
 
 admin.site.register(ThreadedComment, ThreadedCommentsAdmin)
-