+++ threadedcomments/models.py
@@ -1,7 +1,8 @@
-from django.db import models
 from django.conf import settings
+from django.db import models, transaction
 from django.utils.translation import ugettext_lazy as _
-from .compat import Comment, CommentManager, transaction_atomic
+
+from .compat import Comment, CommentManager
 
 PATH_SEPARATOR = getattr(settings, 'COMMENT_PATH_SEPARATOR', '/')
 PATH_DIGITS = getattr(settings, 'COMMENT_PATH_DIGITS', 10)
@@ -9,9 +10,10 @@ PATH_DIGITS = getattr(settings, 'COMMENT_PATH_DIGITS', 10)
 
 class ThreadedComment(Comment):
     title = models.TextField(_('Title'), blank=True)
-    parent = models.ForeignKey('self', null=True, blank=True, default=None, related_name='children', verbose_name=_('Parent'))
+    parent = models.ForeignKey('self', null=True, blank=True, on_delete=models.CASCADE, default=None, related_name='children', verbose_name=_('Parent'))
     last_child = models.ForeignKey('self', null=True, blank=True, on_delete=models.SET_NULL, verbose_name=_('Last child'))
     tree_path = models.CharField(_('Tree path'), max_length=500, editable=False)
+    newest_activity = models.DateTimeField(null=True)
 
     objects = CommentManager()
 
@@ -27,7 +29,7 @@ class ThreadedComment(Comment):
     def root_path(self):
         return ThreadedComment.objects.filter(pk__in=self.tree_path.split(PATH_SEPARATOR)[:-1])
 
-    @transaction_atomic
+    @transaction.atomic
     def save(self, *args, **kwargs):
         skip_tree_path = kwargs.pop('skip_tree_path', False)
         super(ThreadedComment, self).save(*args, **kwargs)
@@ -40,22 +42,31 @@ class ThreadedComment(Comment):
 
             self.parent.last_child = self
             ThreadedComment.objects.filter(pk=self.parent_id).update(last_child=self.id)
+            ThreadedComment.objects.filter(pk=self.parent_id).update(newest_activity=self.submit_date)
+            ThreadedComment.objects.filter(parent_id=self.parent_id).update(newest_activity=self.submit_date)
 
         self.tree_path = tree_path
         ThreadedComment.objects.filter(pk=self.pk).update(tree_path=self.tree_path)
+        ThreadedComment.objects.filter(pk=self.pk).update(newest_activity=self.submit_date)
 
     def delete(self, *args, **kwargs):
         # Fix last child on deletion.
         if self.parent_id:
             try:
-                prev_child_id = ThreadedComment.objects \
-                                .filter(parent=self.parent_id) \
-                                .exclude(pk=self.pk) \
-                                .order_by('-submit_date') \
-                                .values_list('pk', flat=True)[0]
+                prev_child = ThreadedComment.objects \
+                    .filter(parent=self.parent_id) \
+                    .exclude(pk=self.pk) \
+                    .order_by('-submit_date')[0]
             except IndexError:
-                prev_child_id = None
-            ThreadedComment.objects.filter(pk=self.parent_id).update(last_child=prev_child_id)
+                prev_child = None
+            if prev_child:
+                ThreadedComment.objects.filter(pk=self.parent_id).update(last_child=prev_child)
+                ThreadedComment.objects.filter(pk=self.parent_id).update(newest_activity=prev_child.submit_date)
+                ThreadedComment.objects.filter(parent_id=self.parent_id).update(newest_activity=prev_child.submit_date)
+            else:
+                ThreadedComment.objects.filter(pk=self.parent_id).update(last_child=None)
+                ThreadedComment.objects.filter(pk=self.parent_id).update(newest_activity=self.parent.submit_date)
+
         super(ThreadedComment, self).delete(*args, **kwargs)
 
     class Meta(object):