+++ threadedcomments/templatetags/threadedcomments_tags.py
@@ -1,6 +1,8 @@
-import django
 from django import template
+from django.conf import settings
+from django.db.models import Q
 from django.template.loader import render_to_string
+from django.utils.encoding import smart_text
 from threadedcomments.compat import BASE_APP, django_comments as comments
 from threadedcomments.util import annotate_tree_properties, fill_tree as real_fill_tree
 
@@ -14,11 +16,43 @@ else:
 
 register = template.Library()
 
-class BaseThreadedCommentNode(BaseCommentNode):
-    def __init__(self, parent=None, flat=False, root_only=False, **kwargs):
+
+class AdminOverrideCommentNode(BaseCommentNode):
+
+    def get_queryset(self, context):
+        ctype, object_pk = self.get_target_ctype_pk(context)
+        if not object_pk:
+            return self.comment_model.objects.none()
+
+        qs = self.comment_model.objects.filter(
+            content_type=ctype,
+            object_pk=smart_text(object_pk),
+            site__pk=settings.SITE_ID,
+        )
+
+        # The is_public and is_removed fields are implementation details of the
+        # built-in comment model's spam filtering system, so they might not
+        # be present on a custom comment model subclass. If they exist, we
+        # should filter on them.
+        field_names = [f.name for f in self.comment_model._meta.fields]
+        if not self.admin:
+            if 'is_public' in field_names:
+                qs = qs.filter(is_public=True)
+            if getattr(settings, 'COMMENTS_HIDE_REMOVED', True) and 'is_removed' in field_names:
+                qs = qs.filter(is_removed=False)
+
+        return qs
+
+
+class BaseThreadedCommentNode(AdminOverrideCommentNode):
+
+    def __init__(self, parent=None, flat=False, root_only=False, newest=False, limit=False, admin=False, **kwargs):
         self.parent = parent
         self.flat = flat
         self.root_only = root_only
+        self.newest = newest
+        self.limit = limit
+        self.admin = admin
         super(BaseThreadedCommentNode, self).__init__(**kwargs)
 
     @classmethod
@@ -29,7 +63,7 @@ class BaseThreadedCommentNode(BaseCommentNode):
                 raise template.TemplateSyntaxError("Second argument in %r tag must be 'for'" % tokens[0])
 
         extra_kw = {}
-        if tokens[-1] in ('flat', 'root_only'):
+        if tokens[-1] in ('flat', 'root_only', 'newest', 'admin'):
             extra_kw[str(tokens.pop())] = True
 
         if len(tokens) == 5:
@@ -58,27 +92,23 @@ class BaseThreadedCommentNode(BaseCommentNode):
 
     def get_queryset(self, context):
         qs = super(BaseThreadedCommentNode, self).get_queryset(context)
+        if self.limit:
+            parent_qs = qs.exclude(parent__isnull=False).order_by('-newest_activity', '-submit_date').values_list('pk', flat=True)[:self.limit]
+            qs = qs.filter(Q(parent_id__in=parent_qs) | Q(pk__in=parent_qs)).distinct()
         if self.flat:
             qs = qs.order_by('-submit_date')
         elif self.root_only:
             qs = qs.exclude(parent__isnull=False).order_by('-submit_date')
+        elif self.newest:
+            qs = qs.order_by('-newest_activity', 'tree_path')
         return qs
-
-    # For older Django (1.5) versions:
-    def get_query_set(self, context):
-        qs = super(BaseThreadedCommentNode, self).get_query_set(context)
-        if self.flat:
-            qs = qs.order_by('-submit_date')
-        elif self.root_only:
-            qs = qs.exclude(parent__isnull=False).order_by('-submit_date')
-        return qs
-
 
 
 class CommentListNode(BaseThreadedCommentNode):
     """
     Insert a list of comments into the context.
     """
+
     def get_context_value_from_queryset(self, context, qs):
         return list(qs)
 
@@ -87,6 +117,7 @@ class CommentCountNode(CommentListNode):
     """
     Insert a count of comments into the context.
     """
+
     def get_context_value_from_queryset(self, context, qs):
         return qs.count()
 
@@ -151,6 +182,7 @@ class CommentFormNode(BaseThreadedCommentNode):
 
 
 class RenderCommentFormNode(CommentFormNode):
+
     @classmethod
     def handle_token(cls, parser, token):
         """
@@ -197,14 +229,9 @@ class RenderCommentFormNode(CommentFormNode):
                 "comments/%s/form.html" % ctype.app_label,
                 "comments/form.html",
             )
-            context.push()
-            form_str = render_to_string(
-                template_search_list,
-                {"form": self.get_form(context)},
-                context.request,
-            )
-            context.pop()
-            return form_str
+            context = context.flatten()
+            context['form'] = self.get_form(context)
+            return render_to_string(template_search_list, context)
         else:
             return ''
 
@@ -222,7 +249,14 @@ class RenderCommentListNode(CommentListNode):
             raise template.TemplateSyntaxError("Second argument in %r tag must be 'for'" % tokens[0])
 
         extra_kw = {}
-        if tokens[-1] in ('flat', 'root_only'):
+        if tokens[-2] == 'limit':
+            if tokens[-1].isdigit():
+                extra_kw['limit'] = tokens.pop(-1)  # removes limit integer
+                tokens.pop(-1)  # removes 'limit'
+            else:
+                raise template.TemplateSyntaxError(
+                    "When using 'limit' with %r tag, it needs to be followed by a positive integer" % (tokens[0],))
+        if tokens[-1] in ('flat', 'root_only', 'newest', 'admin'):
             extra_kw[str(tokens.pop())] = True
 
         if len(tokens) == 3:
@@ -234,8 +268,8 @@ class RenderCommentListNode(CommentListNode):
         elif len(tokens) == 4:
             # {% render_comment_list for app.models pk %}
             return cls(
-                ctype = BaseCommentNode.lookup_content_type(tokens[2], tokens[0]),
-                object_pk_expr = parser.compile_filter(tokens[3]),
+                ctype=BaseCommentNode.lookup_content_type(tokens[2], tokens[0]),
+                object_pk_expr=parser.compile_filter(tokens[3]),
                 **extra_kw
             )
         else:
@@ -249,20 +283,12 @@ class RenderCommentListNode(CommentListNode):
                 "comments/%s/list.html" % ctype.app_label,
                 "comments/list.html"
             ]
-            # For newer Django (1.6) versions
-            if django.VERSION >= (1,6):
-                qs = self.get_queryset(context)
-            else:
-                qs = self.get_query_set(context)
-            context.push()
-            liststr = render_to_string(
-                template_search_list,
-                {"comment_list": self.get_context_value_from_queryset(
-                    context, qs)},
-                context.request,
-            )
-            context.pop()
-            return liststr
+            qs = self.get_queryset(context)
+            context = context.flatten()
+            context.update({
+                "comment_list": self.get_context_value_from_queryset(context, qs),
+            })
+            return render_to_string(template_search_list, context)
         else:
             return ''