+++ CHANGELOG.rst
@@ -1,3 +1,11 @@
+Version 1.1, 19 februari 2017
+-----------------------------
+
+* Added Django 1.10 support.
+* Added sorting on the latest activity (`{% render_comment_list for [object] newest %}`).
+* Dropped Django 1.5, 1.6 support.
+* Dropped Python 2.6 support.
+
 Version 1.0.1, 17 October 2015
 ------------------------------